🚩: interface 방식과 인덱스 시그니처
